requirements.txt
fastapi #==0.104.1
uvicorn #==0.23.2
sqlalchemy #==2.0.27
python-multipart #==0.0.6
jinja2 #==3.1.2
python-dotenv #==1.0.0
reportlab #==4.0.7
pillow #==10.1.0
httpx #==0.25.2
psutil #==5.9.8
# Supabase and PostgreSQL dependencies
supabase==2.8.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
